Updated test. Added swagger.
