+++ pom.xml
@@ -19,6 +19,7 @@
 		<java.version>1.8</java.version>
 		<log4j2.version>2.8.2</log4j2.version>
 		<apache.commons.version>3.9</apache.commons.version>
+		<springfox.swagger.version>2.6.1</springfox.swagger.version>
 	</properties>
 
 	<dependencies>
@@ -86,6 +87,21 @@
 			<groupId>org.springframework.security</groupId>
 			<artifactId>spring-security-test</artifactId>
 			<scope>test</scope>
+		</dependency>
+
+		<!-- Swagger-->
+		<dependency>
+			<groupId>io.springfox</groupId>
+			<artifactId>springfox-swagger2</artifactId>
+			<version>${springfox.swagger.version}</version>
+			<scope>compile</scope>
+		</dependency>
+
+		<dependency>
+			<groupId>io.springfox</groupId>
+			<artifactId>springfox-swagger-ui</artifactId>
+			<version>${springfox.swagger.version}</version>
+			<scope>compile</scope>
 		</dependency>
 	</dependencies>
