+++ src/test/java/com/documentflow/services/AddressServiceImplTest.java
@@ -1,6 +1,6 @@
 package com.documentflow.services;
 
-import com.documentflow.AbstractDocumentFlowTest;
+import com.documentflow.AbstractContragentTest;
 import com.documentflow.entities.Address;
 import com.documentflow.entities.Contragent;
 import com.documentflow.entities.dto.ContragentDtoAddress;
@@ -15,7 +15,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
-public class AddressServiceImplTest extends AbstractDocumentFlowTest {
+public class AddressServiceImplTest extends AbstractContragentTest {
 
     @Autowired
     private AddressService addressService;
